Add files via upload
Object detection is a computer technology related to computer vision and image processing that detects and defines objects such as humans, buildings and cars from digital images and videos (MATLAB). This technology has the power to classify just one or several objects within a digital image at once. It is based on the fundamental property of feature detection